jax_md/io/io.py moved to jax_md/io.py and few changes
1. Inconsistent array length will be checked by np.hstack
2. Local function name changed
3. Copyright notice updated
4. 1D array will be converted to 2D array (usually species, masses and charges are 1D arrays)
