--- jax_md/io/__init__.py
@@ -1,1 +1,0 @@
-from .io import write_xyz
